+++ org.eclim.core/java/org/eclim/plugin/core/project/ProjectManagement.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2005 - 2012  Eric Van Dewoestine
+ * Copyright (C) 2005 - 2013  Eric Van Dewoestine
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -29,8 +29,6 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.xpath.XPathExpression;
 
 import org.apache.commons.lang.StringUtils;
-
-import org.eclim.Services;
 
 import org.eclim.command.CommandLine;
 import org.eclim.command.Error;
@@ -171,7 +169,7 @@ public class ProjectManagement
     // (IFile.refreshLocal) from kicking off a rebuild workspace job, which in
     // turn can cause issues with pdt select and completion engines
     // (See PhpUtils.waitOnBuild).
-    refresh(project, commandLine);
+    refresh(project, commandLine, false);
 
     for (int ii = 0; ii < natures.size(); ii++){
       ProjectManager manager = getProjectManager((String)natures.get(ii));
@@ -360,14 +358,23 @@ public class ProjectManagement
   public static void refresh(IProject project, CommandLine commandLine)
     throws Exception
   {
+    refresh(project, commandLine, true);
+  }
+
+  private static void refresh(
+      IProject project, CommandLine commandLine, boolean refreshNatures)
+    throws Exception
+  {
     ProjectUtils.assertExists(project);
 
     project.refreshLocal(IResource.DEPTH_INFINITE, null);
 
-    for (String nature : managers.keySet()){
-      if(project.hasNature(nature)){
-        ProjectManager manager = ProjectManagement.getProjectManager(nature);
-        manager.refresh(project, commandLine);
+    if (refreshNatures){
+      for (String nature : managers.keySet()){
+        if(project.hasNature(nature)){
+          ProjectManager manager = ProjectManagement.getProjectManager(nature);
+          manager.refresh(project, commandLine);
+        }
       }
     }