+++ org.eclim.sdt/test/eclim_unit_test_scala/src/eclim/test/complete/TestComplete.scala
@@ -1,0 +1,16 @@
+package eclim.test.complete
+
+import eclim.test.{TestJava, TestScala}
+
+class TestComplete {
+
+  def testScala(){
+    val test = new TestScala
+    test.sc 
+  }
+
+  def testJava(){
+    val test = new TestJava
+    test.j 
+  }
+}