+++ org.eclim.sdt/test/eclim_unit_test_scala/src/eclim/test/search/TestSearch.scala
@@ -1,0 +1,16 @@
+package eclim.test.search
+
+import eclim.test.{TestJava, TestScala}
+
+class TestComplete {
+
+  def testScala(){
+    val test = new TestScala
+    test.scalaMethod1
+  }
+
+  def testJava(){
+    val test = new TestJava
+    test.javaMethod1
+  }
+}